requirements.txt
google-adk>=0.3.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
apscheduler>=3.10.4
